Split loihi files into emulator and hardware
Also some other related moves and renames. No code was changed
aside from updating to new locations and handling issues
related to the new locations.
