+++ nengo_loihi/__init__.py
@@ -2,12 +2,8 @@ import logging
 
 from .version import version as __version__
 
-from . import loihi_api
-from . import loihi_cx
+from . import cx
 from . import builder
-from . import allocators
-from . import loihi_interface
-
 from .simulator import Simulator
 from .config import add_params, set_defaults
