+++ nengo_loihi/hardware/__init__.py
@@ -1,0 +1,2 @@
+from .nxsdk_shim import HAS_NXSDK
+from .simulator import LoihiSimulator
