+++ nengo_loihi/hardware/snips/nengo_io.c.template
