+++ nengo_loihi/hardware/snips/nengo_io.h
