+++ nengo_loihi/hardware/snips/nengo_learn.c
