--- nengo_loihi/tests/test_loihi_cx.py
@@ -1,37 +1,0 @@
-import numpy as np
-
-from nengo_loihi.loihi_cx import CxGroup, CxProbe, CxModel
-
-
-def test_simulator_noise(request, plt, seed):
-    target = request.config.getoption("--target")
-
-    model = CxModel()
-    group = CxGroup(10)
-    group.configure_relu()
-
-    group.bias[:] = np.linspace(0, 0.01, group.n)
-
-    group.enableNoise[:] = 1
-    group.noiseExp0 = -2
-    group.noiseMantOffset0 = 0
-    group.noiseAtDendOrVm = 1
-
-    probe = CxProbe(target=group, key='v')
-    group.add_probe(probe)
-    model.add_group(group)
-
-    model.discretize()
-
-    if target == 'loihi':
-        sim = model.get_loihi(seed=seed)
-        sim.run_steps(1000)
-        y = np.column_stack([
-            p.timeSeries.data for p in sim.board.probe_map[probe]])
-    else:
-        sim = model.get_simulator(seed=seed)
-        sim.run_steps(1000)
-        y = sim.probe_outputs[probe]
-
-    plt.plot(y)
-    plt.yticks(())
